requirements.txt
rawpy>=0.21.0
flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
gunicorn>=22.0.0
